backend/requirements.txt
# Framework Web
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Sécurité et Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# ML 
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2

# Utilitaires
python-dotenv==1.0.0
python-multipart==0.0.6
pyspark==3.5.0
google-genai
argon2-cffi
